# backend/main.py
from fastapi import FastAPI, HTTPException, Depends
from pydantic import BaseModel
from typing import Optional, List
from datetime import datetime

app = FastAPI(title="AI-Powered Book Platform API" )

# Placeholder for user management (e.g., from Better-Auth)
class UserIn(BaseModel):
    username: str
    email: str
    password: str
    programming_background: str
    hardware_owned: str
    gpu_availability: bool
    linux_familiarity: str

class UserOut(BaseModel):
    user_id: str
    username: str
    email: str
    programming_background: str
    hardware_owned: str
    gpu_availability: bool
    linux_familiarity: str
    created_at: datetime

class LoginIn(BaseModel):
    email: str
    password: str

class Token(BaseModel):
    access_token: str
    token_type: str

class ChapterContent(BaseModel):
    chapter_id: str
    title: str
    content_markdown: str

class PersonalizedChapterContent(BaseModel):
    chapter_id: str
    title: str
    personalized_content_markdown: str

class TranslatedChapterContent(BaseModel):
    chapter_id: str
    title: str
    translated_content_markdown: str

class ChatbotQueryIn(BaseModel):
    query_text: str
    chapter_id: str
    selected_text: Optional[str] = None

class ChatbotResponseOut(BaseModel):
    response_text: str
    citations: List[str]

# Placeholder for authentication dependency
def get_current_user():
    # This will be replaced by actual Better-Auth logic
    raise HTTPException(status_code=401, detail="Not authenticated")

@app.post("/auth/signup", status_code=201, response_model=UserOut)
async def signup(user: UserIn):
    # TODO: Implement user registration with Better-Auth and NeonDB
    raise HTTPException(status_code=501, detail="Not implemented")

@app.post("/auth/login", response_model=Token)
async def login(credentials: LoginIn):
    # TODO: Implement user login with Better-Auth
    raise HTTPException(status_code=501, detail="Not implemented")

@app.get("/user/profile", response_model=UserOut)
async def get_user_profile(current_user: UserOut = Depends(get_current_user)):
    # TODO: Implement fetching user profile from NeonDB
    raise HTTPException(status_code=501, detail="Not implemented")

@app.put("/user/profile", response_model=UserOut)
async def update_user_profile(updated_profile: UserIn, current_user: UserOut = Depends(get_current_user)):
    # TODO: Implement updating user profile in NeonDB
    raise HTTPException(status_code=501, detail="Not implemented")

@app.get("/chapters/{chapter_id}", response_model=ChapterContent)
async def get_chapter(chapter_id: str):
    # TODO: Implement fetching chapter content
    raise HTTPException(status_code=501, detail="Not implemented")

@app.post("/chapters/{chapter_id}/personalize", response_model=PersonalizedChapterContent)
async def personalize_chapter(chapter_id: str, current_user: UserOut = Depends(get_current_user), difficulty_level: Optional[str] = None, learning_speed: Optional[str] = None):
    # TODO: Implement personalization using PersonalizerAgent
    raise HTTPException(status_code=501, detail="Not implemented")

@app.post("/chapters/{chapter_id}/translate", response_model=TranslatedChapterContent)
async def translate_chapter(chapter_id: str, target_language: str, current_user: UserOut = Depends(get_current_user)):
    # TODO: Implement translation using TranslatorAgent
    raise HTTPException(status_code=501, detail="Not implemented")

@app.post("/chatbot/query", response_model=ChatbotResponseOut)
async def chatbot_query(query: ChatbotQueryIn, current_user: UserOut = Depends(get_current_user)):
    # TODO: Implement RAG chatbot query using RAGRetrieverAgent
    raise HTTPException(status_code=501, detail="Not implemented")